Change package name and include basic version config.
The orocos_kdl project wasn't able to use the Chocolatey packaged Eigen because
of issues with the package resolution. This renames the package in the CMake registry
to "Eigen" from "Eigen3", which was originally taken from the FindEigen3.cmake module.

It also sets a version config that will accept any version not less than 3.3.3.
By local testing, not specifying minor and patch versions will compare correctly.
That is, requesting version 3 will succeed.