Return package name to Eigen3.
The earlier revisions of the package did use the package name "Eigen3"
which matched the upstream exported package name but Orocos failed
because it looks first for an "Eigen" package. So I changed the name of
the package thinking that standard distributions do not follow
upstream's convention. It was actually orocos that was peculiar not
upstream.

This changes the package name to back to Eigen3 for the benefit of other
packages. This will mean that some consumers that put their find module
onto the path directly will need to have the module modified to check
for a config file package.

For tinyxml_vendor, this involved wrapping the find module logic with a
config package check as below:

```
find_package(TinyXML CONFIG)
if (NOT TinyXML_FOUND)
	# The rest of the find module logic
endif()
```

For any FindEigen3.cmake modules in the ecosystem they should be wrapped
with the conditional below for compatibility with this package.

```
find_package(Eigen3 CONFIG)
if (NOT Eigen3_FOUND)
	message("No Eigen3 package found in CMake's user or system registry.")
	# Existing find module logic.
endif()
```